+++ zenml/steps/sequencer/standard_sequencer/utils.py
@@ -19,12 +19,10 @@ import numpy as np
 import pandas as pd
 import pytz
 
-from zenml.core.standards.standard_keys import MethodKeys, DefaultKeys
-from zenml.core.steps.sequencer.standard_sequencer.methods import \
-    methods_resampling
-from zenml.core.steps.sequencer.standard_sequencer.methods.standard_methods \
-    import \
-    ResamplingMethods
+from zenml.standards.standard_keys import MethodKeys, DefaultKeys
+from zenml.steps.sequencer.standard_sequencer.methods import methods_resampling
+from zenml.steps.sequencer.standard_sequencer.methods.standard_methods \
+    import ResamplingMethods
 
 
 def convert_datetime_to_secs(dt):
